% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{aictab_ind}
\alias{aictab_ind}
\title{Perform model selection over all individuals}
\usage{
aictab_ind(mod_ls, cutoff = 0, K = NULL)
}
\arguments{
\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{A cutoff value to exclude individuals with bad fit, default = -1 indicating model that did not converge will be excluded. Values > 0 will exclude based on coefficient values}
}
\value{
A AIC model selection table
}
\description{
Perform AIC model selection over all individuals by adding up likelihood of individual model (based code partly taken from package AICcmodavg)
}
\examples{
data(goats)
ls1<-list()
ls1[[1]]<-as.formula(STATUS~ELEVATION+SLOPE+ET+ASPECT+HLI+TASP)
ls1[[2]]<-as.formula(STATUS~ET+ASPECT+HLI+TASP)
out<-rsf_ind(goats$ID, data=goats, form_ls=ls1)
aictab_ind(out)
}
