% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{bad_fit1}
\alias{bad_fit1}
\title{Identify potential individual with bad fits based on coefficients and model convergence for a specific model}
\usage{
bad_fit1(mod_ls, cutoff = 1000, m = 1)
}
\arguments{
\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{Value a coeffiecient may take to indicate bad fit (default=1000)}

\item{m}{model number (based on number in list of formula provided to rsf_ind)}
}
\value{
A list with first element giving individuals with bad fits based on coefficients and second element containing individuals with bad fit based on convergence
}
\description{
Identify potential individual with bad fits based on coefficients and model convergence
}
\examples{
Add something
}
