% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{cmodel}
\alias{cmodel}
\title{Remove elements from glm object to save space}
\usage{
cmodel(cm)
}
\arguments{
\item{cm}{A glm object}
}
\value{
A glm object
}
\description{
Remove elements from glm objects (Taken from: https://www.r-bloggers.com/trimming-the-fat-from-glm-models-in-r/)
}
