% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{cmodel_ssf}
\alias{cmodel_ssf}
\title{Remove elements from clogit object to save space}
\usage{
cmodel_ssf(cm)
}
\arguments{
\item{cm}{A coxph object}
}
\value{
A coxph object
}
\description{
Remove elements from glm objects
}
