% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\docType{data}
\name{deer}
\alias{deer}
\title{deer - White-tailed deer data set}
\format{
A data frame with 252500 rows and 8 variables

ID a numeric vector, individuals

STATUS a numeric vector, 1: used, 0: available

step_length a numeric vector of step length (m)

turn_angle a numeric vector of turning angle (in radians)

time the time of the observed or random steps

STRATA each group of observed and random steps

forest % of forest around end location of a step

hummod % human modification around end location of a step
}
\usage{
deer
}
\description{
GPS collar data of 5 white-tailed deer from southern Illinois formatted for a SSF
}
\keyword{datasets}
