% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{eval_ratio}
\alias{eval_ratio}
\title{Evaluate ratio of used and random locations of individuals in a RSF table}
\usage{
eval_ratio(id, value)
}
\arguments{
\item{id}{A vector of individual for each observation}

\item{value}{A vector indicating if each observation is used (=1) or random(=0)}
}
\value{
A list indicating the range in ratio, range in random locations, and range in used location.
}
\description{
Evaluate ratio of used and random locations of individuals in a RSF table
}
