% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\docType{data}
\name{goats}
\alias{goats}
\title{goats - Mountain goats data set}
\format{A data frame with 19014 rows and 8 variables}
\usage{
goats
}
\description{
GPS collar data of mountain goats (Oreamnos americanus) from Lele and Keim (2006).
}
\keyword{datasets}
