% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{ind_coef}
\alias{ind_coef}
\title{Extract individual coefficients}
\usage{
ind_coef(m = 1, mod_ls, cutoff = 0)
}
\arguments{
\item{m}{model number (based on number in list of formula provided to rsf_ind)}

\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{A cutoff value to exclude individuals with bad fit, default = -1 indicating model that did not converge will be excluded. Values > 0 will exclude based on coefficient}
}
\value{
A list containing a table population average with confidence intervals and a table of individual coefficients
}
\description{
Extract individual coefficients.
}
\examples{
Add something
}
