% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{kfold_ind}
\alias{kfold_ind}
\title{Perform kfold cross-validation at the individual level .}
\usage{
kfold_ind(m = 1, mod_ls, ls = ls, cutoff = 0, k = 5, nrepet = 5,
  nbins = 10, grph = T)
}
\arguments{
\item{m}{model number (based on number in list of formula provided to rsf_ind)}

\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{A cutoff value to exclude individuals with bad fit, default = -1 indicating model that did not converge will be excluded. Values > 0 will exclude based on coefficient}

\item{k}{number of fold (default = 5)}

\item{nrepet}{Number of repetitions (default =10)}

\item{nbins}{Number of bins (default =10)}

\item{jitter}{Logical, whether to add some random noise to the predictions (useful when the model is fitted on categorica  variables, which can produces error in the ranking process).}

\item{reproducible}{Logical, whether to use a fixed seed for each repetition.}
}
\value{
A data frame with the correlations (\code{cor}) and the type of value (\code{type}).
}
\description{
Perform kfold cross-validation at the individual level and return histogram, mean kfold accros individual and min/max value
}
