% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{kfold_ind_ssf}
\alias{kfold_ind_ssf}
\title{Perform kfold cross-validation at the individual level for an SSF .}
\usage{
kfold_ind_ssf(m = 1, mod_ls, ls = ls, cutoff = 0, k = 5, nrepet = 10, grph = T)
}
\arguments{
\item{m}{model number (based on number in list of formula provided to rsf_ind)}

\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{ls}{A list of list of formula fed into rsf_ind}

\item{cutoff}{A cutoff value to exclude individuals with bad fit, default = -1 indicating model that did not converge will be excluded. Values > 0 will exclude based on coefficient}

\item{k}{number of fold (default = 5)}

\item{nrepet}{Number of repetitions (default =10)}
}
\value{
A vector of mean kfold score for each individual and an (optional) histogram
}
\description{
Perform kfold cross-validation at the individual level and return histogram and mean kfold accross individual
}
\examples{
data(deer)
deer$log_sl<-log(deer$step_length)
ls2<-list()
ls2[[1]]<-as.formula(STATUS~forest+hummod+log_sl+strata(STRATA))
ls2[[2]]<-as.formula(STATUS~forest+log_sl+strata(STRATA))
out<-ssf_ind(deer$ID, data=deer, form_ls=ls2)
kfold_ind_ssf(m=1, out, ls=ls1)
}
