% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{kfoldRSF}
\alias{kfoldRSF}
\title{Perform kfold cross-validation on a RSF output.}
\usage{
kfoldRSF(
  mod,
  k = 5,
  nrepet = 10,
  nbins = 10,
  jitter = TRUE,
  random = TRUE,
  method = method,
  x = m,
  form_ls = ls,
  reproducible = TRUE
)
}
\arguments{
\item{mod}{A RSF model (glm or glmer)}

\item{k}{number of fold (default = 5)}

\item{nrepet}{Number of repetitions (default =10)}

\item{nbins}{Number of bins (default =10)}

\item{jitter}{Logical, whether to add some random noise to the predictions (useful when the model is fitted on categorica  variables, which can produces error in the ranking process).}

\item{reproducible}{Logical, whether to use a fixed seed for each repetition.}
}
\value{
A data frame with the correlations (\code{cor}) and the type of value (\code{type}).
}
\description{
Perform kfold cross-validation on a RSF output. Similar to what is recommended in Boyce 2002. Function developped with Mathieu Basille
}
