% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{kfoldSSF}
\alias{kfoldSSF}
\title{Cross-validation for regression models.Taken and modified slightly from hab package}
\usage{
kfoldSSF(
  mod,
  k = 5,
  nrepet = 100,
  jitter = TRUE,
  reproducible = TRUE,
  form_ls = ls,
  x = m,
  method = method,
  details = FALSE
)
}
\arguments{
\item{mod}{A fitted model for which there exists a \code{kfold}
method (currently \code{coxph} and \code{clogit} models).}

\item{k}{The number of equal size subsamples of the partition.}

\item{nrepet}{The number of repetitions.}

\item{jitter}{Logical, whether to add some random noise to the
predictions (useful when the model is fitted on categorical
variables, which can produces error in the ranking process).}

\item{reproducible}{Logical, whether to use a fixed seed for each
repetition.}
}
\value{
A data frame with the correlations (\code{cor}) and the
type of value (\code{type}).
}
\description{
Note: needs complete names in the coxph/clogit call, and not
'cos(var)' or 'I(var*10)', except for 'strata()' and
'cluster()'.
}
\author{
Adapted from Mathieu Basille function in hab package
}
