% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{pop_avg}
\alias{pop_avg}
\title{Extract population average of top model and extract individual coefficients}
\usage{
pop_avg(m = 1, mod_ls, cutoff = 0, method = "boot", nboot = 1000)
}
\arguments{
\item{m}{model number (based on number in list of formula provided to rsf_ind)}

\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{A cutoff value to exclude individuals with bad fit, default = -1 indicating model that did not converge will be excluded. Values > 0 will exclude based on coefficient}

\item{method}{If = "boot", population average is based on bootstrap, if = "murtaugh" based on standard errors weighting. See Prokopenko et al 2016 or Murtaugh 2007 for details.}

\item{nboot}{Number of bootstrap iterations, default = 1000. Only applicable if method = "boot".}
}
\value{
A list containing a table population average with confidence intervals and a table of individual coefficients
}
\description{
Extract population average of top model and extract individual coefficients. Population average can be calculated based on bootstrap (Prokopenko et al. 2016 JAppEco) or weighted based on standard errors (Murtaugh 2007 Ecology)
}
\examples{
data(goats)
ls1<-list()
ls1[[1]]<-as.formula(STATUS~ELEVATION+SLOPE+ET+ASPECT+HLI+TASP)
ls1[[2]]<-as.formula(STATUS~ET+ASPECT+HLI+TASP)
out<-rsf_ind(goats$ID, data=goats, form_ls=ls1)
pop_avg(m=1, out, method="murtaugh") 
}
