% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{resample_rsf}
\alias{resample_rsf}
\title{Resample a RSF table to keep constant ratio of used/random locations accross individuals}
\usage{
resample_rsf(data, id = "Id_Year", value = "Value", ratio = 3)
}
\arguments{
\item{data}{The RSF dataset to resample}

\item{id}{A vector of individual for each observation}

\item{value}{A vector indicating if each observation is used (=1) or random(=0)}

\item{ratio}{The ratio of random:used location (default =3, meaning 3 random locations for each used location)}
}
\value{
A RSF dataset
}
\description{
Resample a RSF table to keep constant ratio of used/random locations accross individuals. Resampling is done with replacement.
}
