% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{rm_bad_fit}
\alias{rm_bad_fit}
\title{Remove potential individual with bad fits based on coefficients}
\usage{
rm_bad_fit(mod_ls, cutoff = 1000)
}
\arguments{
\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{Value a coeffiecient may take to indicate bad fit (default=1000)}
}
\value{
A list excluding individual with bad fits.
}
\description{
Remove potential individual with bad fits based on coefficients
}
