% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{rm_bad_fit1}
\alias{rm_bad_fit1}
\title{Remove potential individual with bad fits based on coefficients for a specific model}
\usage{
rm_bad_fit1(mod_ls, cutoff = 1000, m = 1)
}
\arguments{
\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{Value a coeffiecient may take to indicate bad fit (default=1000)}

\item{m}{model number (based on number in list of formula provided to rsf_ind)}
}
\value{
A list excluding individual with bad fits.
}
\description{
Remove potential individual with bad fits based on coefficients
}
