% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{rm_conv_fit}
\alias{rm_conv_fit}
\title{Remove potential individual with bad fits based on model convergence}
\usage{
rm_conv_fit(mod_ls)
}
\arguments{
\item{mod_ls}{A list of list of model generated by rsf_ind}
}
\value{
A list excluding individual with bad fits.
}
\description{
Remove potential individual with bad fits based on model convergence
}
