% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{rm_conv_fit1}
\alias{rm_conv_fit1}
\title{Remove potential individual with bad fits based on model convergence for a specific model}
\usage{
rm_conv_fit1(mod_ls, m = 1)
}
\arguments{
\item{m}{model number (based on number in list of formula provided to rsf_ind)}
}
\value{
A list excluding individual with bad fits.
}
\description{
Remove potential individual with bad fits based on model convergence
}
