% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{rsf_mod}
\alias{rsf_mod}
\title{Apply a list of candidate models to a single individual}
\usage{
rsf_mod(sub, form_ls, cleanModel = F, method = method)
}
\arguments{
\item{sub}{A subset of data from a single individual}

\item{form_ls}{A list of formulas for the different candidate models}

\item{method}{Weither typical glm or bias-reduction glm should be fitted (see package brglm)}

\item{cleamModel}{Whether the model should be "cleaned" to save memory space}
}
\value{
A list of glm objects
}
\description{
Apply a list of candidate models to a single individual
}
