% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{simu_coefs}
\alias{simu_coefs}
\title{Simulate normally-distributed individual coefficients from RSF/SSF based on their standard errors}
\usage{
simu_coefs(coef, se, n = 1000)
}
\arguments{
\item{coef}{A matrix of individual coefficients (output of ind_se)}

\item{n}{Number of random coefficients to generate for each individual (default=1000)}
}
\value{
An array of individual coefficients
}
\description{
Simulate individual coefficients based on standard errors (to propagate uncertainty)
}
\examples{
data(goats)
ls1<-list()
ls1[[1]]<-as.formula(STATUS~ELEVATION+SLOPE+ET+ASPECT+HLI+TASP)
ls1[[2]]<-as.formula(STATUS~ET+ASPECT+HLI+TASP)
out<-rsf_ind(goats$ID, data=goats, form_ls=ls1)
coef<-ind_coef(m=1, out)
se<-ind_se(m1, out)
simu<-simu_coefs(coef, se, n=100)
}
