% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{simu_sd}
\alias{simu_sd}
\title{Individual variation (Heterogeneity) based on simulated coefficients}
\usage{
simu_sd(simu)
}
\arguments{
\item{simu}{An array of simulated individual coefficients based on their uncertainties(output of simu_coefs)}
}
\value{
A matrix of heterogeneity (one value for each covariates and replicates)
}
\description{
Calculate heterogeneity for each replicate of simulated coefficients
}
\examples{
data(goats)
ls1<-list()
ls1[[1]]<-as.formula(STATUS~ELEVATION+SLOPE+ET+ASPECT+HLI+TASP)
out<-rsf_ind(goats$ID, data=goats, form_ls=ls1)
coef<-ind_coef(m=1, out)
se<-ind_se(m=1, out)
simu<-simu_coefs(coef, se, n=100)
sd<-simu_sd(simu)
colnames(sd)<-names(coef)
head(sd)
apply(sd, 2, quantile, na.rm=T) #Show variation around estimate of each covariate
colMeans(sd) #Calculate average heterogeneity for each covariate
}
