% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{simu_spe}
\alias{simu_spe}
\title{Specialization based on simulated coefficients}
\usage{
simu_spe(simu)
}
\arguments{
\item{simu}{An array of simulated individual coefficients based on their uncertainties(output of simu_coefs)}
}
\value{
A matrix of specialization (one value for each covariates and replicates)
}
\description{
Calculate specialization for each replicate of simulated coefficients
}
\examples{
data(goats)
ls1<-list()
ls1[[1]]<-as.formula(STATUS~ELEVATION+SLOPE+ET+ASPECT+HLI+TASP)
out<-rsf_ind(goats$ID, data=goats, form_ls=ls1)
coef<-ind_coef(m=1, out)
se<-ind_se(m=1, out)
simu<-simu_coefs(coef, se, n=100)
spe<-simu_spe(simu)
colnames(spe)<-names(coef)
head(spe)
apply(spe, 2, quantile, na.rm=T) #Show variation around estimate of each covariate
colMeans(spe) #Calculate average specialization for each covariate
}
