% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{ssf_ind}
\alias{ssf_ind}
\title{Apply a list of SSF candidate models to multiple individuals}
\usage{
ssf_ind(id, data, form_ls, cleanModel = T, method = "approximate")
}
\arguments{
\item{id}{A vector indicating the individuals}

\item{data}{The dataset containing all data}

\item{form_ls}{A list of formulas for the different candidate models}

\item{method}{Whether exact or approximate ML should be performed (see package survival)}

\item{cleamModel}{Whether the model should be "cleaned" to save memory space (default = F)}
}
\value{
A list of of coxph objects
}
\description{
Apply ssf_mod to each individual of a dataset
}
