% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{ssf_mod}
\alias{ssf_mod}
\title{Apply a list of candidate SSF models to a single individual}
\usage{
ssf_mod(sub, form_ls, cleanModel = F, method = "approximate")
}
\arguments{
\item{sub}{A subset of data from a single individual}

\item{form_ls}{A list of formulas for the different candidate models}

\item{method}{Whether exact or approximate ML should be performed (see package survival)}

\item{cleamModel}{Whether the model should be "cleaned" to save memory space}
}
\value{
A list of coxph objects
}
\description{
Apply a list of candidate SSF models to a single individual
}
