% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{ind_se}
\alias{ind_se}
\title{Extract individual standard errors}
\usage{
ind_se(m = 1, mod_ls, cutoff = 0, id_year = F)
}
\arguments{
\item{m}{model number (based on number in list of formula provided to rsf_ind)}

\item{mod_ls}{A list of list of model generated by rsf_ind}

\item{cutoff}{A cutoff value to exclude individuals with bad fit, default = -1 indicating model that did not converge will be excluded. Values > 0 will exclude based on coefficient}

\item{id_year}{Whether id_year (instead of individual) are provided. Individual and year needs to be separated by an underscore for the function to work properly.}
}
\value{
A table of individual standard errors for each coefficients
}
\description{
Extract individual standard errors
}
\examples{
data(goats)
ls1<-list()
ls1[[1]]<-as.formula(STATUS~ELEVATION+SLOPE+ET+ASPECT+HLI+TASP)
ls1[[2]]<-as.formula(STATUS~ET+ASPECT+HLI+TASP)
out<-rsf_ind(goats$ID, data=goats, form_ls=ls1)
ind_se(m=1, out) 
}
