% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts RSF.R
\name{rsf_ind}
\alias{rsf_ind}
\title{Apply a list of candidate models to multiple individuals}
\usage{
rsf_ind(id, data, form_ls, cleanModel = F, method = "glm.fit")
}
\arguments{
\item{id}{A vector indicating the individuals}

\item{data}{The dataset containing all data}

\item{form_ls}{A list of formulas for the different candidate models}

\item{method}{Weither typical glm or bias-reduction glm should be fitted (default="glm.fit) (see package brglm)}

\item{cleamModel}{Whether the model should be "cleaned" to save memory space (default = F)}
}
\value{
A list of list of glm objects
}
\description{
Apply rsf_mod to each individual of a dataset
}
\examples{
data(goats)
ls1<-list()
ls1[[1]]<-as.formula(STATUS~ELEVATION+SLOPE+ET+ASPECT+HLI+TASP)
ls1[[2]]<-as.formula(STATUS~ET+ASPECT+HLI+TASP)
out<-rsf_ind(goats$ID, data=goats, form_ls=ls1)
}
